got the waypoints "converted" but not tested, will test with the search algorithm logic machine test

+++ src/behaviours/src/waypoints/Waypoint.h
@@ -11,7 +11,7 @@ class Waypoint : public StateMachine
 
         bool hasArrived( void ) { return has_arrived; }
 
-        std::tuple<float,float> getOutput()
+        std::tuple<int,int> getOutput()
         {
             return std::make_tuple( output_left_pwm, output_right_pwm );
         }
